+++ foxes/output/flow_plots_2d/flow_plots.py
@@ -650,9 +650,12 @@ class FlowPlots2D(Output):
         cmap=None,
         quiver_n=None,
         quiver_pars={},
+        states_sel=None,
+        states_isel=None,
         verbosity=0,
         ret_state=False,
         ret_im=False,
+        ret_data=False,
         animated=False,
         **kwargs,
     ):
@@ -711,12 +714,18 @@ class FlowPlots2D(Output):
             Place a vector at each `n`th point
         quiver_pars: dict, optional
             Parameters for plt.quiver
+        states_sel: list, optional
+            Reduce to selected states
+        states_isel: list, optional
+            Reduce to the selected states indices
         verbosity: int, optional
             The verbosity level
         ret_state: bool
             Flag for state index return
         ret_im: bool
             Flag for image return
+        ret_data: bool
+            Flag for returing image data
         animated: bool
             Switch for usage for an animation
         kwargs: dict, optional
@@ -732,6 +741,9 @@ class FlowPlots2D(Output):
         im: tuple, optional
             The image objects, matplotlib.collections.QuadMesh
             or matplotlib.QuadContourSet
+        im_data: tuple, optional
+            The image data numpy.ndarrays, 
+            (x_pos, y_pos, z_pos, g_pts, data)
 
         """
 
@@ -746,6 +758,8 @@ class FlowPlots2D(Output):
             farm_results=self.fres,
             g_pts=g_pts,
             verbosity=verbosity,
+            sel={FC.STATE: states_sel} if states_sel is not None else None,
+            isel={FC.STATE: states_isel} if states_isel is not None else None,
             **kwargs,
         )
         data = point_results[var].to_numpy()
@@ -763,7 +777,6 @@ class FlowPlots2D(Output):
                 point_results[FV.WS].to_numpy(),
             )
         )
-        del point_results
 
         # find data min max:
         vmin = vmin if vmin is not None else np.nanmin(data)
@@ -778,7 +791,7 @@ class FlowPlots2D(Output):
             y_pos /= normalize_xy
 
         # loop over states:
-        for si, s in enumerate(self.fres[FC.STATE].to_numpy()):
+        for si, s in enumerate(point_results[FC.STATE].to_numpy()):
             if not animated and title is None:
                 ttl = f"State {s}"
                 ttl += f", z =  {int(np.round(z_pos))} m"
@@ -811,8 +824,12 @@ class FlowPlots2D(Output):
                 animated=animated,
             )
 
-            yield out
-
+            if ret_data:
+                out = list(out) if isinstance(out, tuple) else [out]
+                yield tuple(out + [(x_pos, y_pos, z_pos, g_pts[si], data[si])])
+            else:
+                yield out
+        
     def gen_states_fig_xz(
         self,
         var,
@@ -842,9 +859,12 @@ class FlowPlots2D(Output):
         cmap=None,
         quiver_n=None,
         quiver_pars={},
+        states_sel=None,
+        states_isel=None,
         verbosity=0,
         ret_state=False,
         ret_im=False,
+        ret_data=False,
         animated=False,
         **kwargs,
     ):
@@ -907,12 +927,18 @@ class FlowPlots2D(Output):
             Place a vector at ech `n`th point
         quiver_pars: dict, optional
             Parameters for plt.quiver
+        states_sel: list, optional
+            Reduce to selected states
+        states_isel: list, optional
+            Reduce to the selected states indices
         verbosity: int, optional
             The verbosity level
         ret_state: bool
             Flag for state index return
         ret_im: bool
             Flag for image return
+        ret_data: bool
+            Flag for returing image data
         animated: bool
             Switch for usage for an animation
         kwargs: dict, optional
@@ -928,6 +954,9 @@ class FlowPlots2D(Output):
         im: tuple, optional
             The image objects, matplotlib.collections.QuadMesh
             or matplotlib.QuadContourSet
+        im_data: tuple, optional
+            The image data numpy.ndarrays, 
+            (x_pos, y_pos, z_pos, g_pts, data)
 
         """
 
@@ -952,6 +981,8 @@ class FlowPlots2D(Output):
             farm_results=self.fres,
             g_pts=g_pts,
             verbosity=verbosity,
+            sel={FC.STATE: states_sel} if states_sel is not None else None,
+            isel={FC.STATE: states_isel} if states_isel is not None else None,
             **kwargs,
         )
         data = point_results[var].to_numpy()
@@ -1019,7 +1050,11 @@ class FlowPlots2D(Output):
                 animated=animated,
             )
 
-            yield out
+            if ret_data:
+                out = list(out) if isinstance(out, tuple) else [out]
+                yield tuple(out + [(x_pos, y_pos, z_pos, g_pts[si], data[si])])
+            else:
+                yield out
 
     def gen_states_fig_yz(
         self,
@@ -1050,9 +1085,12 @@ class FlowPlots2D(Output):
         cmap=None,
         quiver_n=None,
         quiver_pars={},
+        states_sel=None,
+        states_isel=None,
         verbosity=1,
         ret_state=False,
         ret_im=False,
+        ret_data=False,
         animated=False,
         **kwargs,
     ):
@@ -1115,12 +1153,18 @@ class FlowPlots2D(Output):
             Place a vector at ech `n`th point
         quiver_pars: dict, optional
             Parameters for plt.quiver
+        states_sel: list, optional
+            Reduce to selected states
+        states_isel: list, optional
+            Reduce to the selected states indices
         verbosity: int, optional
             The verbosity level
         ret_state: bool
             Flag for state index return
         ret_im: bool
             Flag for image return
+        ret_data: bool
+            Flag for returing image data
         animated: bool
             Switch for usage for an animation
         kwargs: dict, optional
@@ -1136,6 +1180,9 @@ class FlowPlots2D(Output):
         im: tuple, optional
             The image objects, matplotlib.collections.QuadMesh
             or matplotlib.QuadContourSet
+        im_data: tuple, optional
+            The image data numpy.ndarrays, 
+            (x_pos, y_pos, z_pos, g_pts, data)
 
         """
 
@@ -1160,6 +1207,8 @@ class FlowPlots2D(Output):
             farm_results=self.fres,
             g_pts=g_pts,
             verbosity=verbosity,
+            sel={FC.STATE: states_sel} if states_sel is not None else None,
+            isel={FC.STATE: states_isel} if states_isel is not None else None,
             **kwargs,
         )
         data = point_results[var].to_numpy()
@@ -1223,4 +1272,8 @@ class FlowPlots2D(Output):
                 animated=animated,
             )
 
-            yield out
+            if ret_data:
+                out = list(out) if isinstance(out, tuple) else [out]
+                yield tuple(out + [(x_pos, y_pos, z_pos, g_pts[si], data[si])])
+            else:
+                yield out